IT WORKS
Wow, it actually works. loop for the win!! prints likelihood score for each line, prints sum of each treatment, and prints sum of block.

The one thing that's funky is that it errors out (stdio abort failure) when I use == in the initial conditions statements in main.c (shouldn't happen, but it works with just = so I might ignore and revisit later..)